requirements/requirements-all.txt
python-frontmatter
jinja2>=3.1
logmuse
pyyaml
requests
ubiquerg
